# callbacks: Improve performance of _store_outcomes by ~35x