--- a/ema_workbench/em_framework/callbacks.py
+++ b/ema_workbench/em_framework/callbacks.py
@@ -196,37 +196,26 @@ class DefaultCallback(AbstractCallback):
             log_progress,
         )
 
-        self.cases = None
         self.results = {}
 
-        # determine data types of parameters
-        columns = []
         dtypes = []
-
         for parameter in self.parameters:
-            name = parameter.name
-            dtype = "float"
-
+            dtype = 'float'
             if isinstance(parameter, BooleanParameter):
-                dtype = "bool"
+                dtype = 'bool'
             elif isinstance(parameter, CategoricalParameter):
-                dtype = "object"
+                dtype = 'object'
             elif isinstance(parameter, IntegerParameter):
-                dtype = "int"
-            columns.append(name)
-            dtypes.append(dtype)
+                dtype = 'int'
+            dtypes.append((parameter.name, dtype))
 
-        for name in ["scenario", "policy", "model"]:
-            columns.append(name)
-            dtypes.append("object")
+        dtypes.extend([
+            ("scenario", "object"),
+            ("policy", "object"),
+            ("model", "object"),
+        ])
 
-        index = np.arange(nr_experiments)
-        column_dict = {
-            name: pd.Series(dtype=dtype, index=index) for name, dtype in zip(columns, dtypes)
-        }
-        df = pd.concat(column_dict, axis=1).copy()
-
-        self.cases = df
+        self.cases = np.empty(nr_experiments, dtype=dtypes)
 
         for outcome in self.outcomes:
             shape = outcome.shape
@@ -239,34 +228,24 @@ class DefaultCallback(AbstractCallback):
         policy = experiment.policy
         index = experiment.experiment_id
 
-        self.cases.at[index, "scenario"] = scenario.name
-        self.cases.at[index, "policy"] = policy.name
-        self.cases.at[index, "model"] = experiment.model_name
-
-        for k, v in scenario.items():
-            self.cases.at[index, k] = v
-
-        for k, v in policy.items():
-            self.cases.at[index, k] = v
+        self.cases[index] = (tuple(scenario.values()) + tuple(policy.values()) +
+                             (scenario.name, policy.name, experiment.model_name))
 
     def _store_outcomes(self, case_id, outcomes):
         for outcome in self.outcomes:
-            outcome = outcome.name
-            _logger.debug(f"storing {outcome}")
+            outcome_name = outcome.name
 
             try:
-                outcome_res = outcomes[outcome]
+                outcome_res = outcomes[outcome_name]
             except KeyError:
-                message = f"{outcome} not specified as outcome in " f"model(s)"
+                message = f"{outcome_name} not specified as outcome in model(s)"
                 _logger.debug(message)
             else:
                 try:
-                    self.results[outcome][case_id,] = outcome_res
+                    self.results[outcome_name][case_id,] = outcome_res
                 except KeyError:
                     data = np.asarray(outcome_res)
-
                     shape = data.shape
-
                     if len(shape) > 2:
                         message = self.shape_error_msg.format(len(shape))
                         raise ema_exceptions.EMAError(message)
@@ -274,8 +253,8 @@ class DefaultCallback(AbstractCallback):
                     shape = list(shape)
                     shape.insert(0, self.nr_experiments)
 
-                    self.results[outcome] = self._setup_outcomes_array(shape, data.dtype)
-                    self.results[outcome][case_id,] = outcome_res
+                    self.results[outcome_name] = self._setup_outcomes_array(shape, data.dtype)
+                    self.results[outcome_name][case_id,] = outcome_res
 
     def __call__(self, experiment, outcomes):
         """
@@ -290,11 +269,7 @@ class DefaultCallback(AbstractCallback):
 
         """
         super().__call__(experiment, outcomes)
-
-        # store the case
         self._store_case(experiment)
-
-        # store outcomes
         self._store_outcomes(experiment.experiment_id, outcomes)
 
     def get_results(self):