On a performance profile of the optimized lake model I noticed that the `_store_case` function took up a large portion of the runtime (46.4% of 5000 runs, or 9535ms).

This commit optimizes the `DefaultCallback` class, including its `_store_case` function by:
- Replacing Pandas DataFrames with NumPy structured arrays for faster data access and manipulation.
- Using NumPy functions and data structures where possible.
- Reducing unnecessary function calls.

This results the `_store_case` function only needing 267ms (or 2.3% of the runtime) on 5000 experiment runs, a speedup of ~35x. When experimenting with the `lake_problem` a total speedup of about 60% will be achieved (500 to 800 it/s on my setup). The faster/simpler the model experimented with, the more of this speedup will be noticed.