@EwoutH 2023-03-19
I needed to make a few small adjustments to make the test set pass, mainly in transforming the data.

I don't think this breaks any (important) behaviour, but please review carefully, especially the changes to the test code.