@quaquel 2023-03-21
Not yet. Focus has been on fixing pynetlogo.