@quaquel 2023-03-21
I don't see anything obviously strange with this suggested modification. 

Out of curiosity, why did you choose to use recarrays? Looking at my old code, I see some obvious places for performance increases, even when sticking to pandas.