@EwoutH 2023-03-21
> Out of curiosity, why did you choose to use recarrays?

I tried some different options, those benchmarked the fastest. It was actually suggested by ChatGPT!

While Pandas DataFrames are way more flexible, in general it's also slower than any kind of NumPy array. I think in this case a recarray is a nice middle ground, it's very fast but also keeps the code human readable and interpretable.

The biggest performance win is in initializing the empty array from the start, and filling it a full row at the time. The `.at` lookups in Pandas are also expensive, especially the larger number when iterating through the dictionary.