@quaquel 2023-03-21
I ran a quick benchmark this afternoon focusing on getting rid of the at statements, but trying to set the full row in one go is actually slower than looping, as far as I can tell. See below the code I tested. I want to run a few more tests and do some line profiling to fully understand what is going on here.

Way back when I used recarrays (because pandas did not exist when I started on the workbench) before shifting to pandas a few years back for the 2.x branch of the workbench. 

```python
def _store_case(experiment):
    exp = experiment.scenario.data | experiment.policy.data
    
    exp["scenario"] = experiment.scenario.name
    exp["policy"] = experiment.policy.name
    exp["model"] = experiment.model_name
    
    index = experiment.experiment_id
    
    cases.loc[index] = exp
```