@quaquel 2023-03-21
Ok, I did some research

```python
def _store_case(experiment):
    exp = experiment.scenario.data | experiment.policy.data
    
    exp["scenario"] = experiment.scenario.name
    exp["policy"] = experiment.policy.name
    exp["model"] = experiment.model_name
    
    index = experiment.experiment_id
    
    cases.replace({index:exp})
```

is 50% faster than the current implementation of _store_case. Was based on [this blog post](https://towardsdatascience.com/how-to-boost-pandas-speed-and-process-10m-row-datasets-in-milliseconds-48d5468e269). Will do a test comparing your recarray based implementation with this one next.