@quaquel 2023-03-21
I noticed one potential problem with your implementation

```python
        self.cases[index] = (
            tuple(scenario.values())
            + tuple(policy.values())
            + (scenario.name, policy.name, experiment.model_name)
        )
```

There is no guarantee whatsoever that scenarios.values() will return the values in the same order as expected by the columns. The order of values can even differ from one experiment to the next. To ensure alignment, you must extract the scenario values by looping over the keys in the correct order. Alternatively, the underlying implementation of Scenario and Policy could shift to an OrderedDict, which also ensures a predictable order.