@EwoutH 2023-03-21
Thanks for the careful review, and interesting to see the deep dive!

> There is no guarantee whatsoever that scenarios.values() will return the values in the same order as expected by the columns.

Good catch! Since Python 3.7, dictionaries are [guaranteed to be insertion ordered](https://stackoverflow.com/a/39980744/19370072). Which means as long as we always insert them in this order, we should be fine (maybe we can even use tuples in that case). Are we currently doing that?