@quaquel 2023-03-21
I did not know about the 3.7 change. That is good to know.

That is not a trivial question. I had to go through the code quite a bit to answer this. But yes, a fixed order is used for sampling, and this order is identical to the order of uncertainties/levers that are passed to callback.

However, you can create custom scenarios/policies, and in that case, the order is not guaranteed.