@EwoutH 2023-03-23
Thanks for reviewing. How well is this currently covered by the tests?

I will dive a bit deeper into it.