@quaquel 2023-03-24
It is clear that some tests are missing, so that is the logical next step. 

I do like moving to a more performant setup with the recarray.