@EwoutH 2023-04-09
Awesome, thanks for this! Let me do a final conformation of improved performance and then I will merge.