@EwoutH 2023-04-09
New performance benchmarks. We might lost a little bit here and there, but by far the most of the performance improvement is still there. According to my profiles, it speeds up `_store_outcomes` by about 25x. When running a very simple model, the whole workbench is now over 2.5x as fast.

**Old**
- Profile: Total time 11.172 ms (48.8%), own time 559 ms (2.4%)
- Now runs about 15.000 iterations of the [simple Python model](https://emaworkbench.readthedocs.io/en/latest/basic_tutorial.html#a-simple-model-in-python) per second.

**New**
- Profile: Total time 424 ms (3.7%), own time 90 ms (0.8%).
- Now runs about 40.000 iterations of the [simple Python model](https://emaworkbench.readthedocs.io/en/latest/basic_tutorial.html#a-simple-model-in-python) per second.